test: colocate `decode_with_orientation` tests with implementation
